change type of lat and lng of Store information

+++ ecommerce-api/app/db/migrations/20191025224911_create_store.js
@@ -4,8 +4,8 @@ exports.up = function(knex) {
   return knex.schema.createTable('store', (table) => {
     table.increments('id').primary();
     table.string('storeName');
-    table.float('lat');
-    table.float('lng');
+    table.double('lat');
+    table.double('lng');
     table.string('address');
     table.string('phone');
     table.timestamp('createdAt');
